Move pop_random to junk drawer
